Make the setup-bash script robust & Fixed setup README docs
- [x] Make `setup-bash.sh` robust checking for the `DOTFILES` env
variable and each file. Print helpful error information
- [x] Provide better insturctions for setup with the now independent
`DOTFILES` env variable setup (e.g. we cannot assume anything will be at
`~/.dotfiles/`

+++ bash/setup-bash.sh
@@ -1,4 +1,16 @@
 #!/bin/sh
 
-cp $DOTFILES/bash/.bash_profile ~/.bash_profile
-cp $DOTFILES/bash/.bashrc ~/.bashrc
+: ${DOTFILES?"Need to set DOTFILES environment variable"}
+
+if [ -f $DOTFILES/bash/.bashrc ]; then
+    cp $DOTFILES/bash/.bashrc ~/.bashrc
+else
+    echo "File not found: \$DOTFILES/bash/.bashrc"
+    exit 1
+fi
+if [ -f $DOTFILES/bash/.bash_profile ]; then
+    cp $DOTFILES/bash/.bash_profile ~/.bash_profile
+else
+    exho "File not found: \$DOTFILES/bash/.bash_profile"
+    exit 1
+fi
